feat: WPF dotNET example created
