+++ examples/WPF/WPF dotNET Examples.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.902
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WPF dotNET Examples", "WPF dotNET Examples\WPF dotNET Examples.csproj", "{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Debug|x64.ActiveCfg = Debug|x64
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Debug|x64.Build.0 = Debug|x64
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Debug|x86.ActiveCfg = Debug|x86
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Debug|x86.Build.0 = Debug|x86
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Release|x64.ActiveCfg = Release|x64
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Release|x64.Build.0 = Release|x64
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Release|x86.ActiveCfg = Release|x86
+		{FBAF2D2D-DDEB-4269-AF50-0E10BF6BF82C}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {3AEFEB7D-8EDE-4A5C-B570-A128ACA754C3}
+	EndGlobalSection
+EndGlobal
