+++ examples/WPF/WPF dotNET Examples/App.config
@@ -1,0 +1,14 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+    <startup> 
+        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+    </startup>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="NationalInstruments.Vision.Common" publicKeyToken="18cbae0f9955702a" culture="neutral" />
+        <bindingRedirect oldVersion="0.0.0.0-13.1.35.100" newVersion="13.1.35.100" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
+</configuration>
