+++ examples/WPF/WPF dotNET Examples/App.xaml
@@ -1,0 +1,9 @@
+﻿<Application x:Class="WPF_dotNET_Examples.App"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:local="clr-namespace:WPF_dotNET_Examples"
+             StartupUri="MainWindow.xaml">
+    <Application.Resources>
+         
+    </Application.Resources>
+</Application>
